Copy files using hardlinks when possible
If files are on the same volume. Tested on a linux box, it does seem
to work as expected. Feedbacks from tests on windows appreciated.
Active by default. Also added a `DISABLE_HARDLINKS` option that can
be set to True in the config.py file, in case someone wants to actively
disable hardlinks.